feat: Add --packagename option to make:class command
This commit adds the --packagename option to the make:class command in the Laravel Console Helper package. The --packagename option allows users to specify the package name when creating a new PHP class. If the option is not provided, the default package name is set to 'main'.

This feature enhances the make:class command by providing more flexibility in specifying the package name for the created class.